Move channel cards (Telegram, Slack, Voice, Email) from Settings Channels tab to Contacts page
Introduce ContactSelection enum for type-safe selection in the Contacts tab.
Add an Assistant section to the contacts list that shows channel configuration
when selected. Create AssistantChannelsDetailView with all moved channel cards.
Slim SettingsChannelsTab down to only the Mobile (iOS) card.
