+++ clients/macos/vellum-assistant/Features/Contacts/ContactsListView.swift
@@ -1,14 +1,16 @@
 import SwiftUI
 import VellumAssistantShared
 
-/// Displays the list of contacts, with a guardian section at the top,
-/// search, channel icons, and status indicators.
+/// Displays the list of contacts, with an assistant section at the top,
+/// a guardian ("Me") section, search, channel icons, and status indicators.
 @MainActor
 struct ContactsListView: View {
     @ObservedObject var viewModel: ContactsViewModel
-    @Binding var selectedContactId: String?
+    @Binding var selection: ContactSelection?
+    var store: SettingsStore?
 
     @State private var hoveredContactId: String?
+    @State private var isAssistantHovered = false
 
     var body: some View {
         VStack(alignment: .leading, spacing: VSpacing.lg) {
@@ -46,6 +48,9 @@ struct ContactsListView: View {
 
     private var contactsList: some View {
         VStack(alignment: .leading, spacing: VSpacing.lg) {
+            // Assistant section (always visible, unaffected by search)
+            assistantSection
+
             // Guardian ("Me") section
             if let guardian = viewModel.guardianContact {
                 guardianSection(guardian)
@@ -70,6 +75,71 @@ struct ContactsListView: View {
         }
     }
 
+    // MARK: - Assistant Section
+
+    private var assistantSection: some View {
+        VStack(alignment: .leading, spacing: VSpacing.md) {
+            Text("Assistant")
+                .font(VFont.sectionTitle)
+                .foregroundColor(VColor.textPrimary)
+
+            Button {
+                selection = .assistant
+            } label: {
+                HStack(spacing: VSpacing.md) {
+                    VStack(alignment: .leading, spacing: VSpacing.xs) {
+                        HStack(spacing: VSpacing.sm) {
+                            Text(assistantDisplayName)
+                                .font(VFont.bodyBold)
+                                .foregroundColor(VColor.textPrimary)
+                                .lineLimit(1)
+
+                            Text("Assistant")
+                                .font(VFont.small)
+                                .foregroundColor(.white)
+                                .padding(.horizontal, VSpacing.sm)
+                                .padding(.vertical, VSpacing.xxs)
+                                .background(VColor.accent)
+                                .clipShape(RoundedRectangle(cornerRadius: VRadius.pill))
+                        }
+
+                        Text(assistantChannelSummary)
+                            .font(VFont.caption)
+                            .foregroundColor(VColor.textSecondary)
+                            .lineLimit(1)
+                    }
+
+                    Spacer()
+                }
+                .padding(VSpacing.lg)
+                .frame(maxWidth: .infinity, alignment: .leading)
+                .background(selection == .assistant ? VColor.hoverOverlay.opacity(0.08) : (isAssistantHovered ? VColor.hoverOverlay.opacity(0.04) : Color.clear))
+                .contentShape(Rectangle())
+            }
+            .buttonStyle(.plain)
+            .onHover { hovering in
+                isAssistantHovered = hovering
+            }
+            .vCard(background: VColor.surfaceSubtle)
+        }
+    }
+
+    /// The assistant's display name, loaded from the identity file or falling
+    /// back to "Assistant".
+    private var assistantDisplayName: String {
+        AssistantDisplayName.firstUserFacing(from: [IdentityInfo.load()?.name]) ?? AssistantDisplayName.placeholder
+    }
+
+    /// Summarizes which assistant channels are configured.
+    private var assistantChannelSummary: String {
+        var channels: [String] = []
+        if store?.telegramHasBotToken == true { channels.append("Telegram") }
+        if store?.slackChannelHasBotToken == true && store?.slackChannelHasAppToken == true { channels.append("Slack") }
+        if store?.twilioHasCredentials == true { channels.append("Voice") }
+        if store?.assistantEmail != nil { channels.append("Email") }
+        return channels.isEmpty ? "No channels configured" : channels.joined(separator: ", ")
+    }
+
     // MARK: - Guardian Section
 
     private func guardianSection(_ contact: ContactPayload) -> some View {
@@ -79,7 +149,7 @@ struct ContactsListView: View {
                 .foregroundColor(VColor.textPrimary)
 
             Button {
-                selectedContactId = contact.id
+                selection = .contact(contact.id)
             } label: {
                 HStack(spacing: VSpacing.md) {
                     VStack(alignment: .leading, spacing: VSpacing.xs) {
@@ -112,7 +182,7 @@ struct ContactsListView: View {
                 }
                 .padding(VSpacing.lg)
                 .frame(maxWidth: .infinity, alignment: .leading)
-                .background(selectedContactId == contact.id ? VColor.hoverOverlay.opacity(0.08) : (hoveredContactId == contact.id ? VColor.hoverOverlay.opacity(0.04) : Color.clear))
+                .background(selection == .contact(contact.id) ? VColor.hoverOverlay.opacity(0.08) : (hoveredContactId == contact.id ? VColor.hoverOverlay.opacity(0.04) : Color.clear))
                 .contentShape(Rectangle())
             }
             .buttonStyle(.plain)
@@ -165,7 +235,7 @@ struct ContactsListView: View {
 
     private func contactRow(_ contact: ContactPayload) -> some View {
         Button {
-            selectedContactId = contact.id
+            selection = .contact(contact.id)
         } label: {
             HStack(spacing: VSpacing.md) {
                 VStack(alignment: .leading, spacing: VSpacing.xs) {
@@ -192,7 +262,7 @@ struct ContactsListView: View {
             }
             .padding(VSpacing.lg)
             .frame(maxWidth: .infinity, alignment: .leading)
-            .background(selectedContactId == contact.id ? VColor.hoverOverlay.opacity(0.08) : (hoveredContactId == contact.id ? VColor.hoverOverlay.opacity(0.04) : Color.clear))
+            .background(selection == .contact(contact.id) ? VColor.hoverOverlay.opacity(0.08) : (hoveredContactId == contact.id ? VColor.hoverOverlay.opacity(0.04) : Color.clear))
             .contentShape(Rectangle())
         }
         .buttonStyle(.plain)
@@ -371,7 +441,7 @@ struct ContactsListView: View {
         ScrollView {
             ContactsListView(
                 viewModel: viewModel,
-                selectedContactId: .constant(nil)
+                selection: .constant(.assistant)
             )
             .padding(VSpacing.lg)
         }
